tests/wpt/mozilla/tests/mozilla/bluetooth/getDescriptors/correct-descriptors.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [generic_access.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(generic_access.name))
    .then(service => service.getCharacteristic(device_name.name))
    .then(characteristic => characteristic.getDescriptors(number_of_digitals.name))
    .then(descriptors => {
        for (let i = 1; i < descriptors.length; i++) {
            descriptors[i].readValue()
            .then(value => {
                assert_equals(value.byteLength, 2);
                assert_equals(value.getUint8(0), 11);
            })
        }
    });
}, 'Find descriptors with UUID in service.');
</script>
